[BCW-31] refactor: 인증 로직 분리 및 컴포넌트 구조 정리 - AuthProvider 생성(인증 로직 분리) - RootLayout, App 구조 개선
